Migrate functions computing interactions between geometric shapes
  in a local header so that they can be called from another file.
